+++ crates/ui/widgets/src/dialog/settings_dialog.rs
@@ -20,10 +20,18 @@ pub struct SettingsDialog {
 }
 
 impl SettingsDialog {
-    pub fn new(target_media_type: &TargetMediaType, sub_dir_depth_limit: u8) -> Self {
+    pub fn new(
+        target_media_type: &TargetMediaType,
+        sub_dir_depth_limit: u8,
+        similarity_threshold: f32,
+    ) -> Self {
         Self {
             tab: Tab::default(),
-            general_settings: GeneralSettings::new(target_media_type, sub_dir_depth_limit),
+            general_settings: GeneralSettings::new(
+                target_media_type,
+                sub_dir_depth_limit,
+                similarity_threshold,
+            ),
             ai_settings: AiSettings::default(),
             file_system_settings: FileSystemSettings::default(),
             about: About::default(),