+++ crates/ui/widgets/src/dialog/settings_dialog/tab/ai_settings/message.rs
@@ -2,4 +2,6 @@
 pub enum Message {
     LoadStart,
     Loaded(Option<String>),
+    GetFfmpegStart,
+    FfmpegGot(Option<String>),
 }