+++ crates/ui/widgets/src/dialog/settings_dialog/tab/general_settings.rs
@@ -8,13 +8,19 @@ mod view;
 pub struct GeneralSettings {
     target_media_type: TargetMediaType,
     sub_dir_depth_limit: u8,
+    similarity_threshold: f32,
 }
 
 impl GeneralSettings {
-    pub fn new(target_media_type: &TargetMediaType, sub_dir_depth_limit: u8) -> Self {
+    pub fn new(
+        target_media_type: &TargetMediaType,
+        sub_dir_depth_limit: u8,
+        similarity_threshold: f32,
+    ) -> Self {
         Self {
             target_media_type: target_media_type.to_owned(),
             sub_dir_depth_limit,
+            similarity_threshold,
         }
     }
 }